+++ src/test/java/ch/admin/bj/swiyu/didtoolbox/JCommanderRunnerTest.java
@@ -1,24 +1,29 @@
 package ch.admin.bj.swiyu.didtoolbox;
 
+import ch.admin.bj.swiyu.didtoolbox.context.DidLogCreatorStrategyException;
 import ch.admin.bj.swiyu.didtoolbox.jcommander.AbstractKeyMaterialDidLogCommand;
 import ch.admin.bj.swiyu.didtoolbox.jcommander.CreateDidLogCommand;
 import ch.admin.bj.swiyu.didtoolbox.jcommander.UpdateDidLogCommand;
 import ch.admin.bj.swiyu.didtoolbox.jcommander.VerificationMethodParameters;
+import ch.admin.bj.swiyu.didtoolbox.vc_data_integrity.EdDsaJcs2022VcDataIntegrityCryptographicSuite;
+import ch.admin.bj.swiyu.didtoolbox.vc_data_integrity.VcDataIntegrityCryptographicSuiteException;
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.internal.Console;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URL;
+import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Path;
 import java.security.spec.InvalidKeySpecException;
+import java.util.HashSet;
 import java.util.Set;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * The class was introduced for the sake of being able to test the CLI with no hassle involved.
@@ -30,6 +35,7 @@ class JCommanderRunnerTest extends AbstractUtilTestBase {
     private final JCommander.Builder jCommanderBuilder = JCommander.newBuilder().console(new Console() {
         @Override
         public void print(CharSequence charSequence) {
+            // empty as it's not used
         }
 
         @Override
@@ -56,7 +62,7 @@ class JCommanderRunnerTest extends AbstractUtilTestBase {
     }
 
     @Test
-    void testRunCreateDidLogCommand() {
+    void testRunCreateDidLogCommand(@TempDir Path tmpDir) {
 
         // REMINDER By default, command.methodVersion is set to DidMethodEnum.WEBVH_1_0
         var command = new CreateDidLogCommand();
@@ -74,7 +80,8 @@ class JCommanderRunnerTest extends AbstractUtilTestBase {
             new JCommanderRunner(jCommanderBuilder
                     .addCommand(CreateDidLogCommand.COMMAND_NAME, command)
                     .build(),
-                    UpdateDidLogCommand.COMMAND_NAME
+                    UpdateDidLogCommand.COMMAND_NAME,
+                    tmpDir.toString()
             ).runCreateDidLogCommand(command); // MUT
         });
 
@@ -82,7 +89,7 @@ class JCommanderRunnerTest extends AbstractUtilTestBase {
     }
 
     @Test
-    void testRunCreateDidLogCommandWithKeyPrerotation() {
+    void testRunCreateDidLogCommandWithKeyPrerotation(@TempDir Path tmpDir) {
 
         // REMINDER By default, command.methodVersion is set to DidMethodEnum.WEBVH_1_0
         var command = new CreateDidLogCommand();
@@ -100,7 +107,8 @@ class JCommanderRunnerTest extends AbstractUtilTestBase {
             new JCommanderRunner(jCommanderBuilder
                     .addCommand(CreateDidLogCommand.COMMAND_NAME, command)
                     .build(),
-                    UpdateDidLogCommand.COMMAND_NAME
+                    UpdateDidLogCommand.COMMAND_NAME,
+                    tmpDir.toString()
             ).runCreateDidLogCommand(command); // MUT
         });
 
@@ -185,7 +193,7 @@ class JCommanderRunnerTest extends AbstractUtilTestBase {
         command.jksAlias = "myalias"; // must match one the pre-rotation key(s) set by the initial entry
         // REMINDER Setting command.verifyingKeyPemFiles is optional, but the value MUST match one of the pre-rotation key(s) set by the initial entry
 
-        assertDoesNotThrow(() -> {
+        assertThrowsExactly(JCommanderRunner.CommandException.class, () -> {
             setKeyMaterial(command); // essential
 
             new JCommanderRunner(jCommanderBuilder
@@ -195,7 +203,8 @@ class JCommanderRunnerTest extends AbstractUtilTestBase {
             ).runUpdateDidLogCommand(command); // MUT
         });
 
-        assertFalse(output.isEmpty());
+        // Expect no output, as command failed
+        assertTrue(output.isEmpty());
     }
 
     @Test
@@ -263,5 +272,136 @@ class JCommanderRunnerTest extends AbstractUtilTestBase {
         });
 
         assertFalse(output.isEmpty());
+    }
+
+    @Test
+    void storeKeysOnDisk_noForceOverwrite_success(@TempDir Path tmpDir) throws VcDataIntegrityCryptographicSuiteException {
+        // check that .didtoolbox doesn't exist yet
+        assertFalse(new File( tmpDir.toString() + File.separator + ".didtoolbox").exists());
+
+        var keys = new EdDsaJcs2022VcDataIntegrityCryptographicSuite();
+
+        var runner = new JCommanderRunner(new JCommander(), "", tmpDir.toString());
+        assertDoesNotThrow(() -> runner.storeKeysOnDisk(keys, false));
+
+        assertTrue(new File( tmpDir.toString() + File.separator + ".didtoolbox").exists());
+        var privateKey = new File( tmpDir.toString() + File.separator + ".didtoolbox/id_ed25519");
+        assertTrue(privateKey.exists());
+        var publicKey = new File( tmpDir.toString() + File.separator + ".didtoolbox/id_ed25519.pub");
+        assertTrue(publicKey.exists());
+
+        var cryptoSuite = new EdDsaJcs2022VcDataIntegrityCryptographicSuite(privateKey.toPath());
+        assertEquals(cryptoSuite.getVerificationKeyMultibase(), keys.getVerificationKeyMultibase());
+    }
+
+    @Test
+    void storeKeysOnDisk_noForceOverwriteAlreadyExists_fails(@TempDir Path tmpDir) throws VcDataIntegrityCryptographicSuiteException, FileAlreadyExistsException {
+        // Check that .didtoolbox doesn't exist yet
+        assertFalse(new File( tmpDir.toString() + File.separator + ".didtoolbox").exists());
+
+        var keys = new EdDsaJcs2022VcDataIntegrityCryptographicSuite();
+        var runner = new JCommanderRunner(new JCommander(), "", tmpDir.toString());
+        // Create filled .didtoolbox directory
+        assertDoesNotThrow(() -> runner.storeKeysOnDisk(keys, false));
+
+        // Try to create new directory
+        var newKeys = new EdDsaJcs2022VcDataIntegrityCryptographicSuite();
+        // Returns 1 when something went wrong (files already exists in this case)
+        assertThrowsExactly(JCommanderRunner.CommandException.class, () -> runner.storeKeysOnDisk(newKeys, false));
+
+        var privateKey = new File( tmpDir.toString() + File.separator + ".didtoolbox/id_ed25519");
+        var cryptoSuite = new EdDsaJcs2022VcDataIntegrityCryptographicSuite(privateKey.toPath());
+        assertEquals(cryptoSuite.getVerificationKeyMultibase(), keys.getVerificationKeyMultibase()); // should be old keys
+    }
+
+    @Test
+    void storeKeysOnDisk_withOverwrite_success(@TempDir Path tmpDir) throws VcDataIntegrityCryptographicSuiteException, FileAlreadyExistsException {
+        // Check that .didtoolbox doesn't exist yet
+        assertFalse(new File( tmpDir.toString() + File.separator + ".didtoolbox").exists());
+
+        var keys = new EdDsaJcs2022VcDataIntegrityCryptographicSuite();
+        var runner = new JCommanderRunner(new JCommander(), "", tmpDir.toString());
+        assertDoesNotThrow(() ->runner.storeKeysOnDisk(keys, true));
+
+        var privateKey = new File( tmpDir.toString() + File.separator + ".didtoolbox/id_ed25519");
+        var cryptoSuite = new EdDsaJcs2022VcDataIntegrityCryptographicSuite(privateKey.toPath());
+        assertEquals(cryptoSuite.getVerificationKeyMultibase(), keys.getVerificationKeyMultibase()); // should be old keys
+    }
+
+    @Test
+    void storeKeysOnDisk_withOverwriteAlreadyExisting_success(@TempDir Path tmpDir) throws VcDataIntegrityCryptographicSuiteException, FileAlreadyExistsException {
+        // Check that .didtoolbox doesn't exist yet
+        assertFalse(new File( tmpDir.toString() + File.separator + ".didtoolbox").exists());
+
+        var keys = new EdDsaJcs2022VcDataIntegrityCryptographicSuite();
+        var runner = new JCommanderRunner(new JCommander(), "", tmpDir.toString());
+        // Create filled .didtoolbox directory
+        assertDoesNotThrow(() -> runner.storeKeysOnDisk(keys, false));
+
+        // Try to create new directory
+        var newKeys = new EdDsaJcs2022VcDataIntegrityCryptographicSuite();
+        // Returns 1 when something went wrong (files already exists in this case)
+        assertDoesNotThrow(() -> runner.storeKeysOnDisk(newKeys, true));
+
+        var privateKey = new File( tmpDir.toString() + File.separator + ".didtoolbox/id_ed25519");
+        var cryptoSuite = new EdDsaJcs2022VcDataIntegrityCryptographicSuite(privateKey.toPath());
+        assertEquals(cryptoSuite.getVerificationKeyMultibase(), newKeys.getVerificationKeyMultibase()); // should be old keys
+    }
+
+    @Test
+    void generateAndSaveNewKey_noOverwrite_success(@TempDir Path tmpDir) {
+        // Check that .didtoolbox doesn't exist yet
+        assertFalse(new File( tmpDir.toString() + File.separator + ".didtoolbox").exists());
+
+        var runner = new JCommanderRunner(new JCommander(), "", tmpDir.toString());
+        Set<File> keys = new HashSet<>();
+        assertDoesNotThrow(() -> runner.generateAndSaveNewKey(false, keys));
+        assertEquals(1, keys.size());
+
+        keys.forEach(file -> {
+            assertTrue(file.exists());
+        });
+    }
+
+    @Test
+    void generateAndSaveNewKey_noOverwriteAlreadyExisting_fails(@TempDir Path tmpDir) throws FileAlreadyExistsException, DidLogCreatorStrategyException {
+        // Check that .didtoolbox doesn't exist yet
+        assertFalse(new File( tmpDir.toString() + File.separator + ".didtoolbox").exists());
+
+        var oldKeys = new EdDsaJcs2022VcDataIntegrityCryptographicSuite();
+        var runner = new JCommanderRunner(new JCommander(), "", tmpDir.toString());
+        // Create filled .didtoolbox directory
+        assertDoesNotThrow(() -> runner.storeKeysOnDisk(oldKeys, false));
+
+        Set<File> keys = new HashSet<>();
+        assertThrowsExactly(JCommanderRunner.CommandException.class, () ->runner.generateAndSaveNewKey(false, keys));
+        assertEquals(0, keys.size());
+
+        keys.forEach(file -> {
+            assertTrue(file.exists());
+        });
+    }
+
+    @Test
+    void generateAndSaveNewKey_withOverwrite_success(@TempDir Path tmpDir) throws Exception {
+        // Check that .didtoolbox doesn't exist yet
+        assertFalse(new File( tmpDir.toString() + File.separator + ".didtoolbox").exists());
+
+        var oldKeys = new EdDsaJcs2022VcDataIntegrityCryptographicSuite();
+        var runner = new JCommanderRunner(new JCommander(), "", tmpDir.toString());
+        // Create filled .didtoolbox directory
+        assertDoesNotThrow(() -> runner.storeKeysOnDisk(oldKeys, false));
+
+        Set<File> keys = new HashSet<>();
+        assertDoesNotThrow(() ->runner.generateAndSaveNewKey(true, keys));
+        assertEquals(1, keys.size());
+
+        keys.forEach(file -> {
+            assertTrue(file.exists());
+        });
+
+        var privateKey = new File( tmpDir.toString() + File.separator + ".didtoolbox/id_ed25519");
+        var cryptoSuite = new EdDsaJcs2022VcDataIntegrityCryptographicSuite(privateKey.toPath());
+        assertNotEquals(cryptoSuite.getVerificationKeyMultibase(), oldKeys.getVerificationKeyMultibase()); // keys have changed
     }
 }